pyproject.toml
[project]
name = "talktomycv"
version = "0.1.0"
description = ""
authors = [
    {name = "rahul",email = "rahul2008d@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "openai (>=1.59.9,<2.0.0)",
    "langchain (>=0.3.14,<0.4.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "faiss-cpu (>=1.9.0.post1,<2.0.0)",
    "langchain-community (>=0.3.14,<0.4.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
